stor fix
admindashboard brudt ned i mindre components
audio recorder added
auth store tilføjet
redundante tags fjernet fra index
