+++ package.json
@@ -31,6 +31,7 @@
     "sass": "^1.72.0",
     "v-calendar": "^3.1.2",
     "vue": "^3.4.21",
+    "vue-audio-recorder": "^3.0.1",
     "vue-router": "^4.3.0",
     "vuefire": "^3.1.23"
   },
