+++ src/views/LoginView.vue
@@ -11,42 +11,28 @@
         <input type="password" v-model="password" required>
       </div>
       <button type="submit">Login</button>
-      
     </form>
     <p v-if="errorMessage" class="error">{{ errorMessage }}</p>
   </div>
 </template>
 
 <script setup>
-import { ref } from 'vue';
-import { getAuth, signInWithEmailAndPassword } from "firebase/auth";
-import router from '@/router';
+import { ref, watch } from 'vue';
+import { useAuthStore } from '@/stores/auth';
 
+const { login, clearErrorMessage } = useAuthStore();
 const email = ref('');
 const password = ref('');
-const errorMessage = ref('');
 
 const submitForm = async () => {
-  const auth = getAuth();
-  await login(auth);
+  await login(email.value, password.value);
 };
 
-// Login function
-const login = async (auth) => {
-  try {
-    const userCredential = await signInWithEmailAndPassword(auth, email.value, password.value);
-    const user = userCredential.user;
-    console.log("Admin logged in:", user);
-    router.push("/admin");
-  } catch (error) {
-    errorMessage.value = error.message;
-    console.error("Login error:", error.message);
-  }
-};
-
-
+watch([email, password], () => {
+  clearErrorMessage();
+});
 </script>
 
 <style>
-/* Styles remain the same */
+
 </style>
